docs: add screen recording of using template
Add the screen recording to the README.md file as an image displaying a
screenshot of the recording that links to https://asciinema.org where
the screen recording can be played. The screen recording itself cannot
be embedded in the README.md because GitHub does not allow `script` HTML
tags.

The link cannot be opened in a new tab because GitHub Markdown does not support
it. See https://github.com/mojombo/github-flavored-markdown/issues/28

Using `svg-term-cli` to create an animated SVG does not work, because it
does not play smoothly and does not use the settings applied to the cast
on asciinema.org.

Update dictionary for VS Code workspace to fix spelling errors.

Issues: #11